feat: Move LoadingWrapper component to a components folder
