+++ src/app/me/login/discord/page.tsx
@@ -4,7 +4,7 @@ import { useRouter, useSearchParams } from 'next/navigation';
 import { useEffect, useState } from 'react';
 import { loginDiscord } from '@/lib/apiManager';
 import { httpStatusCodes } from '@/lib/StatusCodes';
-import LoadingWrapper from '../../accounts/connect/Wrapper';
+import LoadingWrapper from '@/components/me/accounts/LoadingWrapper';
 
 const Page = () => {
     const router = useRouter();
